// src/app/popover/interfaces.ts
export interface PopoverProperties {
    event?: MouseEvent;
    targetElement?: HTMLElement;
    element?: HTMLElement;
    popoverText?: string;
    component?: any;
    width?: string;
    height?: string;
    maxWidth?: string;
    minWidth?: string;
    maxHeight?: string;
    left?: number;
    top?: number;
    right?: string;
    bottom?: string;
    zIndex?: number;
    placement?: 'top' | 'bottom' | 'left' | 'right';
    alignToCenter?: boolean;
    overlayBackdrop?: boolean;
    offset?: number;
    theme?: 'dark' | 'light';
    animationDuration?: number;
    animationTimingFunction?: string;
    animationTranslateY?: string;
    animation?: 'fade' | 'upwards';
    popoverClass?: string | string[];
    padding?: string;
    noArrow?: boolean;
    metadata?: any;
    hideDelay?: number;
    showDelay?: number;
    trigger?: 'hover' | 'click';
    whiteSpace?: 'nowrap' | 'normal';
    borderRadius?: string;
    shadow?: boolean;
    pointerEvents?: "auto" | "none";
    fontSize?: string;
    display?: boolean;
}